Separated out methods
Changed left, right, middle pegs to pegs which contains arrays within
it.

Added a method to check if move is legal, first by checking if it is
between 1 and 3, then making sure the destination peg does not already
have a smaller disc.

Added a method to move the disc.

It does not appear to be moving anything, which I think is because I’m
explicitly returning in the check_if_legal method and I believe that is
breaking out of the whole loop. Will need to find a better way to
accomplish the task.